+++ src/rpm/builder.rs
@@ -92,6 +92,9 @@ pub struct RPMBuilder {
     changelog_entries: Vec<String>,
     changelog_times: Vec<u32>,
     compressor: Compressor,
+
+    vendor: Option<String>,
+    repository: Option<String>,
 }
 
 impl RPMBuilder {
@@ -120,7 +123,18 @@ impl RPMBuilder {
             changelog_times: Vec::new(),
             compressor: Compressor::None(Vec::new()),
             directories: BTreeSet::new(),
+            vendor: None,
+            repository: None,
         }
+    }
+
+    pub fn vendor<T: Into<String>>(mut self, content: T) -> Self {
+        self.vendor = Some(content.into());
+        self
+    }
+    pub fn repository<T: Into<String>>(mut self, content: T) -> Self {
+        self.repository = Some(content.into());
+        self
     }
 
     pub fn epoch(mut self, epoch: u32) -> Self {
@@ -948,6 +962,22 @@ impl RPMBuilder {
                 IndexTag::RPMTAG_POSTUN,
                 offset,
                 IndexData::StringTag(post_uninst_script),
+            ));
+        }
+
+        if let Some(vendor) = self.vendor {
+            actual_records.push(IndexEntry::new(
+                IndexTag::RPMTAG_VENDOR,
+                offset,
+                IndexData::StringTag(vendor),
+            ));
+        }
+
+        if let Some(repository) = self.repository {
+            actual_records.push(IndexEntry::new(
+                IndexTag::RPMTAG_URL,
+                offset,
+                IndexData::StringTag(repository),
             ));
         }